fix(payroll): preserve decimal precision in salary history and support gosiSalary request mapping

+++ src/main/java/com/asg/payroll/salarydetails/dto/SalaryHistoryDto.java
@@ -2,6 +2,7 @@ package com.asg.payroll.salarydetails.dto;
 
 import com.asg.common.lib.dto.LovGetListDto;
 import lombok.*;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -13,13 +14,13 @@ public class SalaryHistoryDto {
 
     private Long detRowId;
     private LocalDate incrementDate;
-    private Long basicSalary;
-    private Long gosiSalary;
-    private Long fixedAllowance;
-    private Long fixedOt;
-    private Long hra;
-    private Long transport;
-    private Long gross;
+    private BigDecimal basicSalary;
+    private BigDecimal gosiSalary;
+    private BigDecimal fixedAllowance;
+    private BigDecimal fixedOt;
+    private BigDecimal hra;
+    private BigDecimal transport;
+    private BigDecimal gross;
     private LovGetListDto designationDet;
     private Long noOfTickets;
     private Long ticketPeriod;
